+++ data/property/resolver.go
@@ -1,121 +1,29 @@
 package property
 
 import (
-	"errors"
 	"fmt"
-	"strings"
 
 	"github.com/project-flogo/core/data"
-	"github.com/project-flogo/core/data/coerce"
-	"github.com/project-flogo/core/support/log"
+	"github.com/project-flogo/core/data/resolve"
 )
 
-var (
-	RegisteredResolvers = make(map[string]ExternalResolver)
-	EnabledResolvers    []ExternalResolver
-)
+var propertyResolverInfo = resolve.NewResolverInfo(true, true)
 
-// Resolver used to resolve property value from external configuration like env, file etc
-type ExternalResolver interface {
-	Name() string
-	// Should return value and true if the given key exists in the external configuration otherwise should return nil and false.
-	LookupValue(key string) (interface{}, bool)
+type Resolver struct {
 }
 
-//DEPRECATED
-func RegisterPropertyResolver(resolver ExternalResolver) error {
-	return RegisterExternalResolver(resolver)
+func (*Resolver) GetResolverInfo() *resolve.ResolverInfo {
+	return propertyResolverInfo
 }
 
-func RegisterExternalResolver(resolver ExternalResolver) error {
+//Resolver Property Resolver $property[item]
+func (*Resolver) Resolve(scope data.Scope, item string, field string) (interface{}, error) {
 
-	logger := log.RootLogger()
-
-	resolverName := resolver.Name()
-
-	if resolverName == "" {
-		return fmt.Errorf("an external property resolver must have a non-empty name")
+	manager := DefaultManager()
+	value, exists := manager.GetProperty(item) //should we add the path and reset it to ""
+	if !exists {
+		return nil, fmt.Errorf("failed to resolve Property: '%s', ensure that property is configured in the application", item)
 	}
 
-	if resolver == nil {
-		return fmt.Errorf("cannot register a 'nil' external property resolver")
-	}
-
-	if _, dup := RegisteredResolvers[resolverName]; dup {
-		return fmt.Errorf("external property resolver already registered: %s", resolverName)
-	}
-
-	logger.Debugf("Registering external property resolver [ %s ]", resolverName)
-
-	RegisteredResolvers[resolverName] = resolver
-
-	return nil
-}
-
-func GetExternalPropertyResolver(resolverType string) ExternalResolver {
-	return RegisteredResolvers[resolverType]
-}
-
-func EnableExternalPropertyResolvers(resolverTypes string) error {
-
-	for _, resolverType := range strings.Split(resolverTypes, ",") {
-		resolver := RegisteredResolvers[resolverType]
-		if resolver == nil {
-			errMag := fmt.Sprintf("Unsupported property resolver type - %s. Resolver not registered.", resolverType)
-			return errors.New(errMag)
-		}
-		EnabledResolvers = append(EnabledResolvers, resolver)
-	}
-
-	return nil
-}
-
-func ResolvePropertyExternally(propertyName string) (interface{}, bool) {
-
-	for _, resolver := range EnabledResolvers {
-		// Use resolver
-		value, resolved := resolver.LookupValue(propertyName)
-		if resolved {
-			return value, true
-		}
-	}
-
-	return nil, false
-}
-
-func ExternalResolverProcessor(properties map[string]interface{}) error {
-
-	logger := log.RootLogger()
-
-	var enabledResolvers []string
-	var resolver ExternalResolver
-	for _, resolver = range EnabledResolvers {
-		enabledResolvers = append(enabledResolvers, resolver.Name())
-	}
-	if len(enabledResolvers) == 1 {
-		logger.Infof("Properties will be resolved with the '%s' resolver", EnabledResolvers[0].Name())
-	} else {
-		logger.Infof("Properties will be resolved with these resolvers (in decreasing order of priority): %v", enabledResolvers)
-	}
-
-	for name := range properties {
-		newVal, found := ResolvePropertyExternally(name)
-
-		if !found {
-			logger.Warnf("Property '%s' could not be resolved using property resolver(s). Using default value from flogo.json.", name)
-		} else {
-			// Get datatype of old value
-			dType, _ := data.GetType(properties[name])
-			if dType != data.TypeUnknown {
-				coercedVal, err := coerce.ToType(newVal, dType)
-				if err == nil {
-					properties[name] = coercedVal
-					continue
-				}
-			}
-			properties[name] = newVal
-		}
-	}
-
-	return nil
+	return value, nil
 }
